fixing merge

+++ libft/ft_atoi.c
@@ -6,17 +6,17 @@
 /*   By: thde-sou <thde-sou@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/20 16:07:28 by thde-sou          #+#    #+#             */
-/*   Updated: 2025/12/20 14:54:47 by thde-sou         ###   ########.fr       */
+/*   Updated: 2025/04/20 19:04:04 by thde-sou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-long long	ft_atoi(const char *nptr)
+int	ft_atoi(const char *nptr)
 {
-	int			sign;
-	long long	res;
-	int			a;
+	int	sign;
+	int	res;
+	int	a;
 
 	sign = 1;
 	res = 0;
